Renaming folder to devolute
