+++ README.md
@@ -1,4 +1,4 @@
-# DigiBP Camunda Template
+# devolute Camunda Template
 
 [![License](http://img.shields.io/:license-apache-blue.svg)](http://www.apache.org/licenses/LICENSE-2.0.html)
 [![Deploy to Heroku](https://img.shields.io/badge/deploy%20to-Heroku-6762a6.svg?longCache=true)](https://heroku.com/deploy)
@@ -98,4 +98,4 @@
 - [Andreas Martin](https://github.com/andreasmartin)
 
 ## License
-- [Apache License, Version 2.0](https://github.com/DigiBP/digibp-archetype-camunda-boot/blob/master/LICENSE)
+- [Apache License, Version 2.0](https://github.com/devolute/devolute-archetype-camunda-boot/blob/master/LICENSE)
