+++ app.json
@@ -1,8 +1,8 @@
 {
-  "name": "digibp-camunda-template",
-  "description": "digibp-camunda-template",
-  "repository": "https://github.com/DigiBP/digibp-camunda-template/tree/master",
-  "addons": ["heroku-postgresql:hobby-dev"],
+  "name": "devolute-camunda-template",
+  "description": "devolute-camunda-template",
+  "repository": "https://github.com/omar-devolute/devolute-camunda-template/tree/master",
+  "addons": ["heroku-postgresql:basic"],
   "logo": "https://avatars3.githubusercontent.com/u/25099178?s=200&v=4",
   "keywords": ["FHNW"],
   "env": {
